Comments and clean up
Added short class description comment ot htmlstripper.py

Finalized the MySQLDatabase class to avoid more distractions and changes.
The class now has a function for setting the start value for votes (the
highest value for negative and the lowest value for positive). Updated
comments to be more readable and up-to-date. Also added 'WHERE' clause
 (votes) and columns to 'SELECT' clause.