Text processing
Some questions contain a lot of code examples, and these can contain examples that messes up the conversion from text to HTML/XML, leading to parser fails. Therefore, the code was changed so that on error, None is returned and the given question is removed from the retrieved data set.

For an example of such a question, see:

http://stackoverflow.com/questions/19535331/print-page-specific-area-or-element