mainly backup
Since there was a lot of different testing, the main.py became pretty cluttered and messy. Instead of cleaning it up, I renamed it to old_main to have a backup of all the stuff tried and links to source(s) it was taken from. This file along with the python27 directory will be deleted closer to deadline.

To also avoid to much clutter and mess with the new main.py I'll instead copy in parts in pieces to get a more readable change/commit log (since the new main was created in another file in another project).