Updated to retrieve unprocessed dataset from database
To be able to store a given dataset to file without doing anything to the actual question text (except removing the html), I added a boolean parameter. This way one can store larger amount of data to a file, instead of constantly having to query the database. In addition, it enhances the option for replicability, since this allows for testing the data before and after processing, e.g. removing code samples, removing numeric values and hex, etc.

However, it should be noted its and either or. You either get the pure data, or you get processed data. There is (and most likely won't be) an option to select what parts of the text to clean.