Exhaustive search fitting and training SVM by use of different methods
test_all_algorithms.py is a quick and dirty file written to run comparison of the data before any processing was done (removal of code, hex, numeric, etc), which then tests the following methods: SVC, LinearSVC and SGD. Its written for replication and data output only, and not for it to be a file for review or readability (as-is state).

In the folder test_all_alg_pickle are all the best fit models for the different methods. The name implies what type was used. Short explanation:
's_': stemmed data (after code, etc has beeen removed); start of filename
'u_': unprocessed (original data is used, except for removal of html); start of filename
'cv_': CountVectorizer has been used in this model
'tfid_': TFIDVectorizer has been used in this model
'_lsvc': LinearSVC; end of filename
'_svc': SVC; end of filename
'_sgd': SGD; end of filename