Brilliant error...
As I was testing to see no changes had been made to the training data files, it turned out it had (saw that two of the files had been added to previous commit). As I then re-ran from main.py, I saw that the files were changed yet again. As it turns out, upon loading the unprocessed training set, it overwrote the files containing the cleaned training set. Not knowing how much of the data was altered and at what point, all these has to be re-run again.

This commit is mainly to set test_all_algorithms to use the right filepaths and get the correct training data.