Working on feature detectors - part 5
A lot of code has now been rewritten. This is mainly to balance out new additions, and also to be able to maintain the replicability of the various training that has been done.

Retrieval of training data from database has been changed. It now takes two extra booleans. This way, you can retrieve completely unprocessed data without any feature detectors (and it contains the HTML as-is in database). Or you can create files that contain one of each feature detector for testing (TODO). If both of these are false, training data will be processed as usual, with removal of HTML and adding all feature detectors to the text (TODO: add the new features).

The training file containing unprocessed data has now been changed, due to it originally not containing HTML. As of now, I doubt that this breaks any of the previously created models from "test_all_algorithms". However, it does break future tests, since currently the unprocessed contains HTML, which needs to be removed before it can be tested out (uncertain if I'll bother to fix this).