Working on feature detectors - part 7
Singular feature extraction complete. Still needs to create models and get accuracy for each model, in addition to ensure that the actual training data is updated so that it uses all existing features and is also expandable (e.g. if new features are added later, the only thing needed to do would then be to add call to that function where features are loaded). Shouldn't be that problematic, since separation is already existing in remove_html_tags_from_text