Working on feature detectors - part 10
Added possibility to load other models. Added possibility to create new training datasets, in addition to re-writing some functions. For now, when creating new training data sets, the option for creating feature detectors for words like 'assignment' and words matching those that are listed in the "Tags" table are set to False (ambiguity).

Tags were also originally not compared from those that were attached to questions, due to these being read as string, instead of list (e.g. "<c><multi-threading>". This was fixed by removing the '<' and '>', and adding these to a list. This list then contains a sub-list, where the index is the index of the given question.

It still remains to create models for all singular feature detectors, and also to create a new model for the new training data being created.