Working on feature detectors - updated datasets
The same issue that plagued Tags were also the same issue for 'has_homework'. First of all, the separation between the words 'teach' and 'teaching' was missing, in addition to it replacing words like 'teacher' with 'has_homework er'. There will still be an issue with the plurals of the words, but for now that is not a priority (unless that actually is a part of the vocabulary in the classifier model).