Working on model creation and predictability - Part 1
As long as a model exists, you can now predict which classification it is. For now, it will with the current models only print out class type and not probability. This will be updated when new models are created, since SVC will then have probability=True, enabling the use of "predict_proba()".