All feature detectors model creation and easter egg
To not spend too much time on adjusting and expanding the menu, I simply added a easter egg which creates two additional models for the selected data set. The first is a model using the best settings for the unprocessed ("*_UP_settings.pkl"). The second is a  model trained which uses the same exhaustive search as was used for the unprocessed (no stemming).
If the easter egg isn't "activated", a normal model is created which includes stemming. Note that there must be a model based on the unprocessed version to re-use its best settings.
Another reason, is that this is simply to get more data, and run comparisons. If this were to be a fully functional program, this "easter egg" would not be that relevant for end users.