SGD availability and new default model
Added possibility to create classifier model based on SGD (currently not added to menu, and set to false, when ran it will use SVC). Adjusted SGD to only use 'log', this way one can get access to 'predict_proba'.
The old default model was removed (outdated), and is now replaced with the model created based on the unprocessed data set classifier.